# docs/docs/swagger.yaml
basePath: /
definitions:
...
  internal_transport_http.Response:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 数据载荷
      msg:
        description: 中文提示信息
        type: string
    type: object
...
  internal_transport_http.UpdateSystemConfigRequest:
    properties:
      mailbox:
        $ref: '#/definitions/tempmail_backend_internal_domain.MailboxConfig'
      rateLimit:
        $ref: '#/definitions/tempmail_backend_internal_domain.RateLimitConfig'
      security:
        $ref: '#/definitions/tempmail_backend_internal_domain.SecurityConfig'
      smtp:
        $ref: '#/definitions/tempmail_backend_internal_domain.SMTPConfig'
    type: object
...
  internal_transport_http.attachmentItem:
    properties:
      attachmentId:
        type: string
      contentType:
        type: string
      filename:
        type: string
      size:
        type: integer
    type: object
...
  internal_transport_http.compatMessageListResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/internal_transport_http.messageItem'
        type: array
      nextCursor:
        description: 下一页游标
        type: string
    type: object
  internal_transport_http.configResponse:
    properties:
      domains:
        description: 可用的邮箱域名列表
        items:
          type: string
        type: array
    type: object
...
  internal_transport_http.emailItem:
    properties:
      createdAt:
        type: string
      domain:
        type: string
      email:
        type: string
      emailId:
        type: string
      expiresAt:
        type: string
      name:
        type: string
    type: object
  internal_transport_http.emailListResponse:
    properties:
      emails:
        items:
          $ref: '#/definitions/internal_transport_http.emailItem'
        type: array
      nextCursor:
        description: 下一页游标
        type: string
    type: object
  internal_transport_http.errorResponse:
    properties:
      error:
        type: string
    type: object
  internal_transport_http.generateEmailRequest:
    properties:
      domain:
        description: 邮箱域名
        type: string
      expiryTime:
        description: 过期时间（毫秒）：3600000(1h), 86400000(1d), 604800000(7d), 0(永久)
        type: integer
      name:
        description: 邮箱前缀
        type: string
    type: object
  internal_transport_http.generateEmailResponse:
    properties:
      createdAt:
        description: 创建时间
        type: string
      domain:
        description: 域名
        type: string
      email:
        description: 完整邮箱地址
        type: string
      emailId:
        description: 邮箱ID
        type: string
      expiresAt:
        description: 过期时间
        type: string
      name:
        description: 邮箱前缀
        type: string
      token:
        description: 访问令牌
        type: string
    type: object
...
  internal_transport_http.messageDetailResponse:
    properties:
      attachments:
        items:
          $ref: '#/definitions/internal_transport_http.attachmentItem'
        type: array
      createdAt:
        type: string
      emailId:
        type: string
      from:
        type: string
      html:
        type: string
      isRead:
        type: boolean
      messageId:
        type: string
      subject:
        type: string
      text:
        type: string
      to:
        type: string
    type: object
  internal_transport_http.messageItem:
    properties:
      createdAt:
        type: string
      from:
        type: string
      hasAttachment:
        type: boolean
      isRead:
        type: boolean
      messageId:
        type: string
      subject:
        type: string
    type: object
...
  tempmail_backend_internal_domain.MailboxConfig:
    properties:
      allowedDomains:
        description: 允许的域名列表
        items:
          type: string
        type: array
      defaultTtl:
        description: 默认过期时间，如 "24h"
        type: string
      maxAliases:
        description: 每个邮箱最大别名数
        type: integer
      maxPerIp:
        description: 单IP最大邮箱数
        type: integer
      requireVerification:
        description: 是否需要邮箱验证
        type: boolean
    type: object
...
  tempmail_backend_internal_domain.RateLimitConfig:
    properties:
      burstSize:
        description: 突发请求数
        type: integer
      createMailboxLimit:
        description: 创建邮箱限流（每小时）
        type: integer
      enabled:
        description: 是否启用限流
        type: boolean
      requestsPerMinute:
        description: 每分钟请求数
        type: integer
    type: object
  tempmail_backend_internal_domain.SMTPConfig:
    properties:
      bindAddr:
        description: 监听地址，如 ":25"
        type: string
      domain:
        description: SMTP服务器域名
        type: string
      maxSize:
        description: 最大邮件大小（字节），默认10MB
        type: integer
      readTimeout:
        description: 读取超时（秒），默认60
        type: integer
    type: object
  tempmail_backend_internal_domain.SecurityConfig:
    properties:
      enableCaptcha:
        description: 是否启用验证码
        type: boolean
      jwtAccessExpiry:
        description: JWT访问令牌有效期，如 "15m"
        type: string
      jwtRefreshExpiry:
        description: JWT刷新令牌有效期，如 "7d"
        type: string
      maxLoginAttempts:
        description: 最大登录尝试次数
        type: integer
      passwordMinLength:
        description: 最小密码长度
        type: integer
    type: object
  tempmail_backend_internal_domain.SystemConfig:
    properties:
      id:
        type: string
      mailbox:
        $ref: '#/definitions/tempmail_backend_internal_domain.MailboxConfig'
      rateLimit:
        $ref: '#/definitions/tempmail_backend_internal_domain.RateLimitConfig'
      security:
        $ref: '#/definitions/tempmail_backend_internal_domain.SecurityConfig'
      smtp:
        $ref: '#/definitions/tempmail_backend_internal_domain.SMTPConfig'
      updatedAt:
        type: string
      updatedBy:
        description: 更新者用户ID
        type: string
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: TempMail 后端 API 文档
  title: TempMail Backend API
  version: 0.9.0
paths:
  /api/config:
    get:
      description: 获取系统配置，包括可用的邮箱域名列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http.configResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取系统配置
      tags:
      - Compat
  /api/emails:
    get:
      description: 获取用户的邮箱列表，支持分页
      parameters:
      - description: 分页游标
        in: query
        name: cursor
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http.emailListResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取邮箱列表
      tags:
      - Compat
  /api/emails/{emailId}:
    get:
      description: 获取指定邮箱的邮件列表，支持分页
      parameters:
      - description: 邮箱ID
        in: path
        name: emailId
        required: true
        type: string
      - description: 分页游标
        in: query
        name: cursor
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http.compatMessageListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取邮件列表
      tags:
      - Compat
  /api/emails/{emailId}/{messageId}:
    get:
      description: 获取邮件的详细内容
      parameters:
      - description: 邮箱ID
        in: path
        name: emailId
        required: true
        type: string
      - description: 邮件ID
        in: path
        name: messageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http.messageDetailResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取单封邮件
      tags:
      - Compat
  /api/emails/generate:
    post:
      consumes:
      - application/json
      description: 创建一个新的临时邮箱
      parameters:
      - description: 邮箱参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.generateEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http.generateEmailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: 生成临时邮箱
      tags:
      - Compat
  /v1/admin/config:
    get:
      description: 获取当前系统配置（需要管理员权限）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_transport_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/tempmail_backend_internal_domain.SystemConfig'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      summary: 获取系统配置
      tags:
      - Admin - Config
    put:
      consumes:
      - application/json
      description: 更新系统配置（需要超级管理员权限）
      parameters:
      - description: 配置信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.UpdateSystemConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_transport_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/tempmail_backend_internal_domain.SystemConfig'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      summary: 更新系统配置
      tags:
      - Admin - Config
  /v1/admin/config/reset:
    post:
      description: 将系统配置重置为默认值（需要超级管理员权限）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_transport_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/tempmail_backend_internal_domain.SystemConfig'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      summary: 重置系统配置
      tags:
      - Admin - Config
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 兼容层 API Key
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 使用格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
  MailboxToken:
    description: 邮箱访问令牌
    in: header
    name: X-Mailbox-Token
    type: apiKey
swagger: "2.0"